Fix exploit in /nick command (#71)
* Fix exploit in /nick command

The nick command currently allows an exploit (CVE-1957-XKCD) that allows an arbitrary user to break the Code of Conduct (which shall be referred to as the CoC from here on out) to invoke the name of [redacted].
This commit kicks the player trying to do so as a warning and changes their nick to a SWF CoC approved version.

* Fix explicit language in NickCommand.java

Unfortunately, I did not follow the correct spelling according to our Lord and Savior Orpheus. I have fixed this grave mistake in this PR. May she forgive me in time.